Add documentation for editor color scheme plugin installtion
JetBrains announced a new feature for the platform that allows to export
editor color schemes as plugins (1) which can be easily installed from the
"JetBrains Plugins Repository" (2). Color scheme plugins can be filtered
with the "Editor Color Schemes" (3) category.

This feature greatly improves the installation process for Nord regarding
the UX and version update handling compared to the manual installation
and is the recommended method.

References:

  (1) https://blog.jetbrains.com/platform/2017/12/export-intellij-editor-themes-as-plugins
  (2) https://plugins.jetbrains.com
  (3) https://plugins.jetbrains.com/tag?headline=151-editor-color-schemes

GH-28