Use opacity of 10% for "File Colors" UI theme keys
The "File Colors" (1) are used to distingush between project files,
folders or packages of specific scopes. Unfortunately the theme API
doesn't allow (yet) to specify the target for the color like e.g.
background or only foreground.
With the previous colors (100% opacity) this resulted in too bright
highlighted backgrounds making most labels unreadable and destroying
the themes overall appearance.

They color's opacity has now been changed to 10% which is still enough
to be recognizable and doesn't affects the readability of labels.

References:
  (1) https://www.jetbrains.com/help/idea/settings-file-colors.html

GH-47