Use nord3 as background color for editor notifiations
The editor color scheme provides the `NOTIFICATION_BACKGROUND` key that
allows to style the background color of notifications from the editor
that are placed right below the tab bar at the top of the editor.
Previously `nord4` was used that made the color almost unreadable since
is uses a lower contrast by default instead of pure black.
This has now been changed to use `nord3` making the actual content
readable again.

GH-52