Fix unreadable completion popup list items in non-focused state (#89)
In version 0.8.0 (1) some new UI theme keys have been added and
deprecated ones modified or removed through GH-86 and PR GH-87.
This included the `ui.CompletionPopup.nonFocusedMask` key that has been
set to `nord0` without a alpha layer value. Since the auto completion
popup of IntelliSense if not focused but the editor itself, the overlay
mask gets applied. As soon as the up or down arrow keys are pressed the
popup gets the focus and therefore the mask gets removed showing the
list items with correct styles.

The key has been removed, the popup should be shown the same like in
focused state for better readable text and easy recognition of the
desired auto complete entry.

References:
  (1) https://github.com/arcticicestudio/nord-jetbrains/releases/tag/v0.8.0

Fixes GH-88