Fix invisible background color of inactive completion popup item
Before this commit, the background color of the marked entry in the
completion popup when triggered while typing
(enabled "Show suggestions as you type" option in preferences) [1]
was the same like the background of the popup itself. As a result,
the user did not know which entry was selected to complete the current
line.
The problem only occured when the completion popup was triggered while
typing but not when being explicitly invoked using the configured
keymapping [2] (`Ctrl` / `⌃` + `Space` by default).
This was because when the popup is invoked while typing, the matching
entry in the list is "inactive", but the color for inative entries was
the same like the background color of the popup itself.

This problem is now fixed by changing the color of the co#rresponding UI
theme key `ui.CompletionPopup.selectionInactiveBackground` from `nord1`
to `nord3`.

[1]: https://www.jetbrains.com/help/idea/auto-completing-code.html#configure-code-completion
[2]: https://www.jetbrains.com/help/idea/auto-completing-code.html#invoke-basic-completion

Related to GH-94
GH-95