Restore removed `parent_scheme` attribute (#137)
In GH-117, the `parent_scheme` attribute has been removed since it was
suspected that it caused "random highlighting breakages" that are
documented in detail in GH-120.

Anyway, contrary to the presumption that the attribute is not required,
it caused syntax elements of many different languages to ignore some
colors defined by the Nord plugin, like e.g. "markup" elements in
(documentation) comments, strings in PHP (GH-119), data flow control
characters like braces in TypeScript/JavaScript and even UI elements
like tabs. See feedback comments of @Tom1206 [1] and @yuru7 [2] in
GH-120 for more examples.

The color `#808080` was used for all these elements instead which is
"hardcoded" in different places in the IDE core platform code [3].
It was not possible to fix these elements using the available editor
scheme keys.

By simply adding back the attribute with the value `Darcula` all these
elements will "magically" use the colors defined by Nord again instead
of `#808080`. It is a strange behavior that this attribute is required
for almost no reason, but it has been added back again to fix the
massive style problems occurred as of Nord plugin version 0.10.0 [4] in
combination with the latest IDE versions 2019.3.3 [5] (that was release
the same day like the plugin update...)

[1]: https://github.com/Tom1206
[2]: https://github.com/yuru7
[3]: https://github.com/JetBrains/intellij-community/search?q=808080&unscoped_q=808080
[4]: https://github.com/arcticicestudio/nord-jetbrains/releases/tag/v0.10.0
[5]: https://blog.jetbrains.com/idea/2020/02/intellij-idea-2019-3-3-is-out

Fixes GH-129, GH-119