Fix editor & breadcrumbs separator line color
Fixed the color of the editor & breadcrumbs separator line to use the
correct color `nord3` instead of the pinkish "test color" that should
only be used during development to quickly identify UI and syntax
elements that are affected by a specific theme color key.

Thanks to "Tom1206" again for the quick bug report.

[1]: https://github.com/Tom1206

Related to GH-125, GH-136
GH-138